requirements.txt
ruamel.yaml>=0.17.17
beautifulsoup4>=4.10.0
requests>=2.25.1